requirements.txt
asgiref==3.2.3
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.0
uritemplate==3.0.1
